flask/requirements.txt
flask
pandas
requests
sqlalchemy
sqlalchemy-utils
api-offres-emploi
psycopg2